backend/requirements_dev.txt
appnope==0.1.2
asgiref==3.4.1
attrs==20.3.0
autopep8==1.5.7
backcall==0.2.0
backports.entry-points-selectable==1.1.1
certifi==2020.12.5
cffi==1.15.0
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
coveralls==1.11.1
cryptography==35.0.0
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.3
dj-rest-auth==2.1.11
Django==3.2
django-allauth==0.46.0
django-environ==0.4.5
django-extensions==3.1.3
django-factory-boy==1.0.0
django-redis==4.12.1
-e git+https://github.com/mkurnikov/django-stubs@861c6653fda9a65d8b2fdc8873ff0f5d0a03839d#egg=django_stubs
django-stubs-ext==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
docopt==0.6.2
drf-yasg==1.20.0
factory-boy==3.2.1
Faker==9.5.2
filelock==3.3.2
flake8==3.9.1
flake8-commas==2.0.0
flake8-mypy==17.8.0
gevent==21.1.2
greenlet==1.0.0
gunicorn==20.1.0
identify==2.3.5
idna==3.1
inflection==0.5.1
iniconfig==1.1.1
ipython==7.29.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy==0.711
mypy-extensions==0.4.3
nodeenv==1.6.0
oauthlib==3.1.1
packaging==21.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==2.4.0
pluggy==1.0.0
pre-commit==2.15.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py==1.11.0
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
Pygments==2.8.1
PyJWT==2.3.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0
raven==6.10.0
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
responses==0.12.1
rope==0.19.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
sentry-sdk==1.0.0
simplegeneric==0.8.1
six==1.15.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
typed-ast==1.4.3
types-pytz==2021.3.0
types-PyYAML==6.0.0
typing-extensions==3.10.0.2
unittest-xml-reporting==3.0.4
uritemplate==4.1.1
urllib3==1.26.4
virtualenv==20.10.0
wcwidth==0.2.5
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.4.0
